some bugfixes;

+++ t2/ast.h
@@ -5,11 +5,11 @@ typedef struct node TreeNode;
 struct node
 {
 	int node_type;
-	int value;
+	double value;
 	TreeNode* left;
 	TreeNode* right;
 };
 
-TreeNode *create_node(int t, int val, TreeNode *l, TreeNode *r);
+TreeNode *create_node(int t, double val, TreeNode *l, TreeNode *r);
 
 void RPN_Walk(TreeNode* aux);
